Moved all files into appropriate folders to avoid name overlapping or confusion with other people's classes.
