+++ src/main/java/controllers/ChatRoomControllers/InitializeViewInterface.java
@@ -1,4 +1,4 @@
-package controllers;
+package controllers.ChatRoomControllers;
 
 public interface InitializeViewInterface {
     UpdateViewPresenterInterface initializeView(SendMessageController messenger);
