+++ src/main/java/controllers/ChatRoomControllers/SendMessageInterface.java
@@ -1,4 +1,4 @@
-package controllers;
+package controllers.ChatRoomControllers;
 
 import entities.User;
